+++ service/history/mutableStateBuilder_test.go
@@ -568,7 +568,7 @@ func (s *mutableStateSuite) prepareTransientWorkflowTaskCompletionFirstBatchRepl
 	workflowTimeout := 222 * time.Second
 	runTimeout := 111 * time.Second
 	workflowTaskTimeout := 11 * time.Second
-	workflowTaskAttempt := int64(1)
+	workflowTaskAttempt := int32(1)
 
 	eventID := int64(1)
 	workflowStartEvent := &historypb.HistoryEvent{
@@ -662,7 +662,7 @@ func (s *mutableStateSuite) prepareTransientWorkflowTaskCompletionFirstBatchRepl
 	err = s.msBuilder.ReplicateWorkflowTaskFailedEvent()
 	s.Nil(err)
 
-	workflowTaskAttempt = int64(123)
+	workflowTaskAttempt = int32(123)
 	newWorkflowTaskScheduleEvent := &historypb.HistoryEvent{
 		Version:   version,
 		EventId:   eventID,
